requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.2
openai==2.15.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stripe==12.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
Pillow
